+++ Main/UI/Elements/Slots/SlotContainerType.cs
@@ -1,8 +1,0 @@
-﻿namespace VMFramework.UI
-{
-    public enum SlotContainerType
-    {
-        This,
-        Background
-    }
-}